End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { MessageParams } from '../_models/messageParams';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,7 @@ export class MessagesComponent implements OnInit {
   messageParams: MessageParams;
   loading = false;
 
-  constructor(private messageService: MessageService) {
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) {
     this.messageParams = this.messageService.getMessageParams();
   }
 
@@ -34,8 +35,12 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Confirm Message Deletion', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
   }
